Refactor - split out TaskQueue from Thread
