+++ CMakeLists.txt
@@ -5,7 +5,9 @@ option(Threads_BuildTests "Build the unit tests when BUILD_TESTING is enabled." 
 
 set(SOURCES
 	"include/Threads/Signal.hpp"
-	"include/Threads/Thread.hpp")
+	"include/Threads/Thread.hpp"
+    "include/Threads/TaskQueue.hpp"
+)
 
 if (${CMAKE_VERSION} VERSION_GREATER "3.19.0")
 	add_library(${PROJECT_NAME} INTERFACE ${SOURCES})
