+++ Tests/TaskQueueTests.hpp
@@ -1,0 +1,14 @@
+//
+//  TaskQueueTests.hpp
+//  Threads
+//
+//  Created by Gusts Kaksis on 11/11/2022.
+//  Copyright © 2022 Gusts Kaksis. All rights reserved.
+//
+
+#ifndef TaskQueueTests_hpp
+#define TaskQueueTests_hpp
+
+void runTaskQueueTests();
+
+#endif /* TaskQueueTests_hpp */
